added stop condistion on regression trees. mse evaluation and testing on boston housing data set. remove stipulation that a feature can only be used for a split once - I think that is a legacy from before I switched to one-hot encoding

+++ mlreview/utils/featurizing.py
@@ -40,7 +40,7 @@ class DictVectorizer:
         if isinstance(X, dict):
             X = [X]
 
-        data = np.zeros((len(X), len(self.feature_to_index)))
+        data = np.zeros((len(X), len(self.feature_to_index)), dtype=int)
         for i, x in enumerate(X):
             for feature, val in x.items():
                 if isinstance(val, str):
